LibWeb/Bindings: Relocate title_case_to_snake_case to utils
