+++ Meta/CMake/libweb_generators.cmake
@@ -299,6 +299,7 @@ function (generate_js_bindings target)
     set(bindings_generator_dependencies
         "${bindings_generator}"
         "${LADYBIRD_SOURCE_DIR}/Meta/Utils/lexer.py"
+        "${LADYBIRD_SOURCE_DIR}/Meta/Utils/utils.py"
         "${LADYBIRD_SOURCE_DIR}/Meta/Utils/webidl_parser.py")
 
     set(exposed_interface_sources
